dynamic_ensemble_rl_trading/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
torch>=1.10.0
torchvision>=0.11.0
xgboost>=1.5.0
scikit-learn>=1.0.0

# Reinforcement Learning
gymnasium>=0.26.0
stable-baselines3>=1.6.0

# Technical Analysis
pandas-ta>=0.3.14b
TA-Lib>=0.4.24

# Data Processing
Pillow>=8.3.0
opencv-python>=4.5.0

# Visualization
matplotlib>=3.4.0
plotly>=5.0.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0

# Documentation
sphinx>=4.0.0

